improve smart on fhir handling logic
ResourceAuthorizationInterceptor.java allow calls for metadata to go through regardless of caller credentials. OAuth validation still takes place but no role specific checks take place. This is require to handle the cases from hapi fhir client library, which makes a single call to metadata endpoint before performing any other calls. This can be disabled on the client side but we cannot rely on that on the server side. Also added dedicated rules for patient resource for the smart on fhir case.